% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_iv.R
\name{bcf_iv}
\alias{bcf_iv}
\title{Bayesian Causal Forest with Instrumental Variables}
\usage{
bcf_iv(
  y,
  w,
  z,
  x,
  binary = FALSE,
  n_burn = 500,
  n_sim = 500,
  inference_ratio = 0.5,
  max_depth = 2,
  cp = 0.01,
  minsplit = 10,
  adj_method = "holm",
  seed = 42
)
}
\arguments{
\item{y}{outcome vector (n x 1).}

\item{w}{vector of treatment received  (n x 1).}

\item{z}{vector of treatment assigned (i.e., instrumental variable)  (n x 1).}

\item{x}{covariates matrix (n x p).}

\item{binary}{Boolean to identify whether the outcome is binary or not, i.e.,
continuous/discrete (default: FALSE).}

\item{n_burn}{burn-in MCMC iterations for Bayesian Causal Forest
(default: 500).}

\item{n_sim}{iterations to save post burn-in for Bayesian Causal Forest
(default: 500).}

\item{inference_ratio}{\% of observations to be assigned to the inference
subsample (default: 0.5).}

\item{max_depth}{maximal depth for the generated CART (default: 2).}

\item{cp}{complexity parameter for the generated CART (default: 0.01).}

\item{minsplit}{minimum observations needed to perform a binary split in the
tree (default: 10).}

\item{adj_method}{p-value adjustment method (default: "holm"), other
options are "bonferroni", "hockberg", "hommel", "BH", "BY", "fdr", "none".}

\item{seed}{random seed for reproducible results (default: 42).}
}
\value{
List with 8 elements:
\itemize{
\item a tree structure discovering the heterogeneity in the causal effects,
\item TSLS estimates of the Complier Average Causal Effects (CCACE) within its nodes,
\item p-values for each CCACE,
\item adjusted p-value for each CCACE,
\item Intention-to-Treat within all the nodes,
\item proportion of Compliers within all the nodes,
\item p-value for weak-iv test,
\item proportion of observations in the node.
}
}
\description{
Function for discovery and estimation of heterogeneity in the Complier
Average Causal Effect in Instrumental Variable settings.
}
