% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate Dataset for BCF-IV Example}
\usage{
generate_dataset(
  n = 1000,
  p = 10,
  rho = 0,
  null = 0,
  effect_size = 2,
  compliance = 0.75
)
}
\arguments{
\item{n}{number of data points (default: 1000)}

\item{p}{number of covariates (default: 10)}

\item{rho}{correlation within the covariates (default: 0)}

\item{null}{effect size for null condition (default: 0)}

\item{compliance}{compliance rate (default: 0.75)}

\item{seq}{effect size (default: 2)}
}
\value{
A list containing the different variables in the generated dataset (y,z,w,X).
}
\description{
Function for generating a dataset for the discovery and estimation of heterogeneity
in the Complier Average Causal Effect in Instrumental Variable settings.
}
\examples{
dataset <- generate_dataset()

}
