% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcf_itt.R
\name{bcf_itt}
\alias{bcf_itt}
\title{Bayesian Causal Forest with Intention To Treat}
\usage{
bcf_itt(
  y,
  w,
  z,
  x,
  binary = FALSE,
  max_depth = 2,
  n_burn = 500,
  n_sim = 500,
  inference_ratio = 0.5,
  seed = 42
)
}
\arguments{
\item{y}{outcome vector (n x 1).}

\item{w}{vector of treatment received  (n x 1).}

\item{z}{vector of treatment assigned (i.e., instrumental variable)  (n x 1).}

\item{x}{covariates matrix (n x p).}

\item{binary}{Boolean to identify whether the outcome is binary or not, i.e.,
continuous/discrete (default: FALSE).}

\item{max_depth}{maximal depth for the generated CART (default: 2).}

\item{n_burn}{burn-in MCMC iterations for Bayesian Causal Forest (default:
500).}

\item{n_sim}{iterations to save post burn-in for Bayesian Causal Forest
(default: 500).}

\item{inference_ratio}{\% of observations to be assigned to the inference
subsample (default: 0.5).}

\item{seed}{random seed for reproducible results (default: 42).}
}
\value{
List with 4 elements:
\itemize{
\item a tree structure discovering the heterogeneity in the causal effects,
\item TSLS estimates of the Complier Average Causal Effects (CCACE) within its nodes,
\item p-values for each CCACE,
\item p-value for weak-iv test.
}
}
\description{
Function for discovery and estimation of heterogeneity in the
Intention-To-Treat in Instrumental Variable settings.
}
